requirements.txt
google-cloud
google-oauth2-tool
google-api-python-client
google-cloud-bigquery
google-cloud-storage
pandas
scikit-learn
xgboost
db-dtypes
fastapi
python-multipart
streamlit
